http2: revert more of upstream http2 change
Even the part we previously kept was bad. Revert all of 82780d60 but
keep 6826f5a7db (which depended on bits of 82780d60). So this is a mix.

TestTransportRetryAfterRefusedStream fails with this change, but only
because it was adjusted in 82780d60 to pass with 82780d60, and this test
doesn't revert all the test changes. I just skip that test instead, because
it doesn't really affect us.

Updates tailscale/corp#12296
Updates golang/go#60818

Signed-off-by: Brad Fitzpatrick <bradfitz@tailscale.com>